fix(daemon): retry the same SCAN, not just the same run; extract the runner
Review round 1, both findings.

RED — "same run" was not "same scan". The mode derives from the run
number AND the mutable watermark, and a partially-successful scan moves
the watermark: a fresh node's run-0 seedFromCursor that seeds blocks
0..1999 and then fails would, recomputed, become a seedFull (run 0 +
now-seeded watermark) — unbounded, from block zero, page budget gone.
The failed attempt's options are now captured and the retry reuses
them; seedFromCursor resumes from the saved cursor, which is what
"retry" ought to mean. Cache released on success.

YELLOW — the runner and its mode policy move to
daemon/chain-discovery-scan.ts. Narrow dependency surface (two agent
methods + a log sink); retry/cadence changes no longer edit the
3.8k-line lifecycle module. The test imports from the new module.

Mutation: reverting the retry to recompute-from-watermark fails the new
partial-seed test with exactly the reviewed failure (call 2 = seedFull).
